+++ tests/ProdManager/models/test_models_appconfig.py
@@ -4,7 +4,7 @@ from ProdManager import create_app
 
 from ProdManager.models.AppConfig import AppConfig
 
-app = create_app()
+app = create_app(scheduled_jobs=False)
 
 def test_app_config():
   with app.app_context():