+++ tests/ProdManager/routes/maintenance/test_routes_maintenance_views.py
@@ -12,7 +12,7 @@ from ProdManager import create_app
 class TestRoutesMaintenanceViews(flask_unittest.AppTestCase):
 
   def create_app(self):
-    app = create_app()
+    app = create_app(scheduled_jobs=False)
     app.config['WTF_CSRF_ENABLED'] = False
     return app